Factor out tc.c functions into headers.
