+++ bpf-gpl/jump.h
@@ -57,6 +57,11 @@ CALI_MAP_V1(cali_v4_state,
 		__u32, struct cali_tc_state,
 		1, 0, MAP_PIN_GLOBAL)
 
+static CALI_BPF_INLINE struct cali_tc_state *state_get(void)
+{
+	__u32 key = 0;
+	return cali_v4_state_lookup_elem(&key);
+}
 
 struct bpf_map_def_extended __attribute__((section("maps"))) cali_jump = {
 	.type = BPF_MAP_TYPE_PROG_ARRAY,
