+++ bpf-gpl/tc.c
@@ -41,335 +41,13 @@
 #include "reasons.h"
 #include "icmp.h"
 #include "arp.h"
+#include "fib.h"
+#include "tc.h"
+#include "policy_program.h"
 
-#ifndef CALI_FIB_LOOKUP_ENABLED
-#define CALI_FIB_LOOKUP_ENABLED true
-#endif
-
-#ifndef CALI_DROP_WORKLOAD_TO_HOST
-#define CALI_DROP_WORKLOAD_TO_HOST false
-#endif
-
-#ifdef CALI_DEBUG_ALLOW_ALL
-
-/* If we want to just compile the code without defining any policies and to
- * avoid compiling out code paths that are not reachable if traffic is denied,
- * we can compile it with allow all
+/* calico_tc is the main function used in all of the tc programs.  It is specialised
+ * for particular hook at build time based on the CALI_F build flags.
  */
-static CALI_BPF_INLINE enum calico_policy_result execute_policy_norm(struct __sk_buff *skb,
-				__u8 ip_proto, __u32 saddr, __u32 daddr, __u16 sport, __u16 dport)
-{
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wunused-label"
-
-	RULE_START(0);
-	RULE_END(0, allow);
-
-	return CALI_POL_NO_MATCH;
-deny:
-	return CALI_POL_DENY;
-allow:
-	return CALI_POL_ALLOW;
-#pragma clang diagnostic pop
-}
-#else
-
-static CALI_BPF_INLINE enum calico_policy_result execute_policy_norm(struct __sk_buff *skb,
-				__u8 ip_proto, __u32 saddr, __u32 daddr, __u16 sport, __u16 dport)
-{
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wunused-label"
-
-	RULE_START(0);
-	RULE_END(0, deny);
-
-	return CALI_POL_NO_MATCH;
-deny:
-	return CALI_POL_DENY;
-allow:
-	return CALI_POL_ALLOW;
-#pragma clang diagnostic pop
-}
-
-#endif /* CALI_DEBUG_ALLOW_ALL */
-
-static CALI_BPF_INLINE struct cali_tc_state * state_get(void)
-{
-	__u32 key = 0;
-	return cali_v4_state_lookup_elem(&key);
-}
-
-__attribute__((section("1/0")))
-int calico_tc_norm_pol_tail(struct __sk_buff *skb)
-{
-	CALI_DEBUG("Entering normal policy tail call\n");
-
-	struct cali_tc_state *state = state_get();
-	if (!state) {
-	        CALI_DEBUG("State map lookup failed: DROP\n");
-	        goto deny;
-	}
-
-	state->pol_rc = execute_policy_norm(skb, state->ip_proto, state->ip_src,
-					    state->ip_dst, state->sport, state->dport);
-
-	bpf_tail_call(skb, &cali_jump, EPILOGUE_PROG_INDEX);
-	CALI_DEBUG("Tail call to post-policy program failed: DROP\n");
-
-deny:
-	return TC_ACT_SHOT;
-}
-
-struct fwd {
-	int res;
-	__u32 mark;
-	enum calico_reason reason;
-#if FIB_ENABLED
-	__u32 fib_flags;
-	bool fib;
-#endif
-};
-
-#if FIB_ENABLED
-#define fwd_fib(fwd)			((fwd)->fib)
-#define fwd_fib_set(fwd, v)		((fwd)->fib = v)
-#define fwd_fib_set_flags(fwd, flags)	((fwd)->fib_flags = flags)
-#else
-#define fwd_fib(fwd)	false
-#define fwd_fib_set(fwd, v)
-#define fwd_fib_set_flags(fwd, flags)
-#endif
-
-static CALI_BPF_INLINE struct fwd calico_tc_skb_accepted(struct __sk_buff *skb,
-							 struct iphdr *ip_header,
-							 struct cali_tc_state *state,
-							 struct calico_nat_dest *nat_dest);
-
-static CALI_BPF_INLINE int skb_nat_l4_csum_ipv4(struct __sk_buff *skb, size_t off,
-						__be32 ip_from, __be32 ip_to,
-						__u16 port_from, __u16 port_to,
-						__u64 flags)
-{
-	int ret = 0;
-
-	if (ip_from != ip_to) {
-		CALI_DEBUG("L4 checksum update (csum is at %d) IP from %x to %x\n", off,
-				bpf_ntohl(ip_from), bpf_ntohl(ip_to));
-		ret = bpf_l4_csum_replace(skb, off, ip_from, ip_to, flags | BPF_F_PSEUDO_HDR | 4);
-		CALI_DEBUG("bpf_l4_csum_replace(IP): %d\n", ret);
-	}
-	if (port_from != port_to) {
-		CALI_DEBUG("L4 checksum update (csum is at %d) port from %d to %d\n",
-				off, bpf_ntohs(port_from), bpf_ntohs(port_to));
-		int rc = bpf_l4_csum_replace(skb, off, port_from, port_to, flags | 2);
-		CALI_DEBUG("bpf_l4_csum_replace(port): %d\n", rc);
-		ret |= rc;
-	}
-
-	return ret;
-}
-
-static CALI_BPF_INLINE int forward_or_drop(struct __sk_buff *skb,
-					   struct cali_tc_state *state,
-					   struct fwd *fwd)
-{
-	int rc = fwd->res;
-	enum calico_reason reason = fwd->reason;
-
-	if (rc == TC_ACT_SHOT) {
-		goto deny;
-	}
-
-	if (rc == CALI_RES_REDIR_BACK) {
-		int redir_flags = 0;
-		if  (CALI_F_FROM_HOST) {
-			redir_flags = BPF_F_INGRESS;
-		}
-
-		/* Revalidate the access to the packet */
-		if ((void *)(long)skb->data + sizeof(struct ethhdr) > (void *)(long)skb->data_end) {
-			reason = CALI_REASON_SHORT;
-			goto deny;
-		}
-
-		/* Swap the MACs as we are turning it back */
-		struct ethhdr *eth_hdr = (void *)(long)skb->data;
-		unsigned char mac[ETH_ALEN];
-		__builtin_memcpy(mac, &eth_hdr->h_dest, ETH_ALEN);
-		__builtin_memcpy(&eth_hdr->h_dest, &eth_hdr->h_source, ETH_ALEN);
-		__builtin_memcpy(&eth_hdr->h_source, mac, ETH_ALEN);
-
-		rc = bpf_redirect(skb->ifindex, redir_flags);
-		if (rc == TC_ACT_REDIRECT) {
-			CALI_DEBUG("Redirect to the same interface (%d) succeeded.\n", skb->ifindex);
-			goto skip_fib;
-		}
-
-		CALI_DEBUG("Redirect to the same interface (%d) failed.\n", skb->ifindex);
-		goto deny;
-	} else if (rc == CALI_RES_REDIR_IFINDEX) {
-		struct arp_value *arpv;
-		__u32 iface = state->ct_result.ifindex_fwd;
-
-		struct arp_key arpk = {
-			.ip = state->ip_dst,
-			.ifindex = iface,
-		};
-
-		arpv = cali_v4_arp_lookup_elem(&arpk);
-		if (!arpv) {
-			CALI_DEBUG("ARP lookup failed for %x dev %d\n",
-					bpf_ntohl(state->ip_dst), iface);
-			goto skip_redir_ifindex;
-		}
-
-		/* Revalidate the access to the packet */
-		if ((void *)(long)skb->data + sizeof(struct ethhdr) > (void *)(long)skb->data_end) {
-			reason = CALI_REASON_SHORT;
-			goto deny;
-		}
-		/* Patch in the MAC addresses that should be set on the next hop. */
-		struct ethhdr *eth_hdr = (void *)(long)skb->data;
-		__builtin_memcpy(&eth_hdr->h_dest, arpv->mac_dst, ETH_ALEN);
-		__builtin_memcpy(&eth_hdr->h_source, arpv->mac_src, ETH_ALEN);
-
-		rc = bpf_redirect(iface, 0);
-		if (rc == TC_ACT_REDIRECT) {
-			CALI_DEBUG("Redirect directly to interface (%d) succeeded.\n", iface);
-			goto skip_fib;
-		}
-
-skip_redir_ifindex:
-		CALI_DEBUG("Redirect directly to interface (%d) failed.\n", iface);
-		/* fall through to FIB if enabled or the IP stack, don't give up yet. */
-	}
-
-#if FIB_ENABLED
-	// Try a short-circuit FIB lookup.
-	if (fwd_fib(fwd)) {
-		/* XXX we might include the tot_len in the fwd, set it once when
-		 * we get the ip_header the first time and only adjust the value
-		 * when we modify the packet - to avoid geting the header here
-		 * again - it is simpler though.
-		 */
-		if (skb_too_short(skb)) {
-			reason = CALI_REASON_SHORT;
-			CALI_DEBUG("Too short\n");
-			goto deny;
-		}
-
-		struct iphdr *ip_header = skb_iphdr(skb);
-		struct bpf_fib_lookup fib_params = {
-			.family = 2, /* AF_INET */
-			.tot_len = bpf_ntohs(ip_header->tot_len),
-			.ifindex = skb->ingress_ifindex,
-			.l4_protocol = state->ip_proto,
-			.sport = bpf_htons(state->sport),
-			.dport = bpf_htons(state->dport),
-		};
-
-		/* set the ipv4 here, otherwise the ipv4/6 unions do not get
-		 * zeroed properly
-		 */
-		fib_params.ipv4_src = state->ip_src;
-		fib_params.ipv4_dst = state->ip_dst;
-
-		CALI_DEBUG("FIB family=%d\n", fib_params.family);
-		CALI_DEBUG("FIB tot_len=%d\n", fib_params.tot_len);
-		CALI_DEBUG("FIB ifindex=%d\n", fib_params.ifindex);
-		CALI_DEBUG("FIB l4_protocol=%d\n", fib_params.l4_protocol);
-		CALI_DEBUG("FIB sport=%d\n", bpf_ntohs(fib_params.sport));
-		CALI_DEBUG("FIB dport=%d\n", bpf_ntohs(fib_params.dport));
-		CALI_DEBUG("FIB ipv4_src=%x\n", bpf_ntohl(fib_params.ipv4_src));
-		CALI_DEBUG("FIB ipv4_dst=%x\n", bpf_ntohl(fib_params.ipv4_dst));
-
-		CALI_DEBUG("Traffic is towards the host namespace, doing Linux FIB lookup\n");
-		rc = bpf_fib_lookup(skb, &fib_params, sizeof(fib_params), fwd->fib_flags);
-		if (rc == 0) {
-			CALI_DEBUG("FIB lookup succeeded\n");
-
-			/* Since we are going to short circuit the IP stack on
-			 * forward, check if TTL is still alive. If not, let the
-			 * IP stack handle it. It was approved by policy, so it
-			 * is safe.
-			 */
-			if ip_ttl_exceeded(ip_header) {
-				rc = TC_ACT_UNSPEC;
-				goto cancel_fib;
-			}
-
-			// Update the MACs.  NAT may have invalidated pointer into the packet so need to
-			// revalidate.
-			if ((void *)(long)skb->data + sizeof(struct ethhdr) > (void *)(long)skb->data_end) {
-				reason = CALI_REASON_SHORT;
-				goto deny;
-			}
-			struct ethhdr *eth_hdr = (void *)(long)skb->data;
-			__builtin_memcpy(&eth_hdr->h_source, fib_params.smac, sizeof(eth_hdr->h_source));
-			__builtin_memcpy(&eth_hdr->h_dest, fib_params.dmac, sizeof(eth_hdr->h_dest));
-
-			// Redirect the packet.
-			CALI_DEBUG("Got Linux FIB hit, redirecting to iface %d.\n", fib_params.ifindex);
-			rc = bpf_redirect(fib_params.ifindex, 0);
-			/* now we know we will bypass IP stack and ip->ttl > 1, decrement it! */
-			if (rc == TC_ACT_REDIRECT) {
-				ip_dec_ttl(ip_header);
-			}
-		} else if (rc < 0) {
-			CALI_DEBUG("FIB lookup failed (bad input): %d.\n", rc);
-			rc = TC_ACT_UNSPEC;
-		} else {
-			CALI_DEBUG("FIB lookup failed (FIB problem): %d.\n", rc);
-			rc = TC_ACT_UNSPEC;
-		}
-	}
-
-cancel_fib:
-#endif /* FIB_ENABLED */
-
-skip_fib:
-
-	if (CALI_F_TO_HOST) {
-		/* If we received the packet from the tunnel and we forward it to a
-		 * workload we need to skip RPF check since there might be a better path
-		 * for the packet if the host has multiple ifaces and might get dropped.
-		 *
-		 * XXX We should check ourselves that we got our tunnel packets only from
-		 * XXX those devices where we expect them before we even decap.
-		 */
-		if (CALI_F_FROM_HEP && state->tun_ip != 0) {
-			fwd->mark = CALI_SKB_MARK_SKIP_RPF;
-		}
-		/* Packet is towards host namespace, mark it so that downstream
-		 * programs know that they're not the first to see the packet.
-		 */
-		CALI_DEBUG("Traffic is towards host namespace, marking with %x.\n", fwd->mark);
-		/* FIXME: this ignores the mask that we should be using.
-		 * However, if we mask off the bits, then clang spots that it
-		 * can do a 16-bit store instead of a 32-bit load/modify/store,
-		 * which trips up the validator.
-		 */
-		skb->mark = fwd->mark | CALI_SKB_MARK_SEEN; /* make sure that each pkt has SEEN mark */
-	}
-
-	if (CALI_LOG_LEVEL >= CALI_LOG_LEVEL_INFO) {
-		__u64 prog_end_time = bpf_ktime_get_ns();
-		CALI_INFO("Final result=ALLOW (%d). Program execution time: %lluns\n",
-				rc, prog_end_time-state->prog_start_time);
-	}
-
-	return rc;
-
-deny:
-	if (CALI_LOG_LEVEL >= CALI_LOG_LEVEL_INFO) {
-		__u64 prog_end_time = bpf_ktime_get_ns();
-		CALI_INFO("Final result=DENY (%x). Program execution time: %lluns\n",
-				reason, prog_end_time-state->prog_start_time);
-	}
-
-	return TC_ACT_SHOT;
-}
-
 static CALI_BPF_INLINE int calico_tc(struct __sk_buff *skb)
 {
 	struct cali_tc_state *state;
